+++ src/lib/keyboard_layouts/keyboard_layout_win_fr.py
